requirements/base.txt
python-dotenv==0.10.1


graphene==2.1.3
tornado==5.1.1
graphene-tornado==2.0.1
traitlets==4.3.1

dnspython==1.16.0
motor==2.0